+++ TP4_Montecarlo/WorkerSocket.java
@@ -34,7 +34,7 @@ public class WorkerSocket {
 		// compute
 		System.out.println("TODO : compute Monte Carlo and send total");
             Master m = new Master();
-            int iterations = Integer.parseInt(str);
+            long iterations = Long.parseLong(str);
             long totalCount = m.doRun(iterations, 1, "test.txt");
 	        pWrite.println(totalCount);         // send number of points in quarter of disk
 	    }else{
